+++ hls/lib/hls_endpoint.ex
@@ -18,6 +18,7 @@ defmodule Membrane.RTC.Engine.Endpoint.HLS do
   alias Membrane.RTC.Engine
   alias Membrane.RTC.Engine.Endpoint.HLS.{HLSConfig, MixerConfig}
   alias Membrane.RTC.Engine.Endpoint.WebRTC.TrackReceiver
+  alias Membrane.RTC.Engine.Subscriber
   alias Membrane.RTC.Engine.Track
   alias Membrane.Time
 
@@ -121,16 +122,23 @@ defmodule Membrane.RTC.Engine.Endpoint.HLS do
   end
 
   @impl true
-  def handle_init(_context, options) do
+  def handle_init(ctx, options) do
+    {:endpoint, endpoint_id} = ctx.name
+
+    subscriber = %Subscriber{
+      subscribe_mode: options.subscribe_mode,
+      endpoint_id: endpoint_id,
+      rtc_engine: options.rtc_engine
+    }
+
     state =
       options
       |> Map.from_struct()
       |> Map.merge(%{
-        tracks: %{},
         stream_beginning: nil,
         terminating?: false,
         start_mixing_sent?: false,
-        subscribed_endpoints: MapSet.new()
+        subscriber: subscriber
       })
 
     {[notify_parent: :ready], state}
@@ -156,11 +164,14 @@ defmodule Membrane.RTC.Engine.Endpoint.HLS do
       |> Enum.map(&{&1, track_id})
       |> Enum.filter(&Map.has_key?(ctx.children, &1))
 
-    {removed_track, tracks} = Map.pop!(state.tracks, track_id)
-    state = %{state | tracks: tracks}
+    {removed_track, subscriber} = Subscriber.pop_track!(state.subscriber, track_id)
+
+    state = %{state | subscriber: subscriber}
 
     sink_bin_used? =
-      Enum.any?(tracks, fn {_id, track} ->
+      state.subscriber
+      |> Subscriber.get_tracks()
+      |> Enum.any?(fn {_id, track} ->
         track.stream_id == removed_track.stream_id
       end)
 
@@ -189,7 +200,7 @@ defmodule Membrane.RTC.Engine.Endpoint.HLS do
       ) do
     {offset, state} = get_track_offset(state)
 
-    track = Map.get(state.tracks, track_id)
+    track = Subscriber.get_track(state.subscriber, track_id)
     track_spec = get_track_spec(offset, bin_input(pad), track, state)
 
     {spec, state} =
@@ -260,86 +271,20 @@ defmodule Membrane.RTC.Engine.Endpoint.HLS do
   end
 
   @impl true
-  def handle_parent_notification({:new_tracks, tracks}, ctx, %{subscribe_mode: :auto} = state) do
-    subscribed_tracks =
-      tracks
-      |> Enum.map(fn track -> track.id end)
-      |> subscribe_for_tracks(ctx, state)
+  def handle_parent_notification({:new_tracks, tracks}, _ctx, state) do
+    subscriber = Subscriber.handle_new_tracks(tracks, state.subscriber)
 
-    {[], add_tracks(tracks, subscribed_tracks, state)}
-  end
-
-  @impl true
-  def handle_parent_notification({:new_tracks, tracks}, ctx, %{subscribe_mode: :manual} = state) do
-    subscribed_tracks = Map.keys(state.tracks)
-
-    subscribed_tracks =
-      tracks
-      |> Enum.filter(fn track ->
-        MapSet.member?(state.subscribed_endpoints, track.origin) and
-          track.id not in subscribed_tracks
-      end)
-      |> Enum.map(fn track -> track.id end)
-      |> subscribe_for_tracks(ctx, state)
-
-    {[], add_tracks(tracks, subscribed_tracks, state)}
+    {[], %{state | subscriber: subscriber}}
   end
 
   @impl true
   def handle_parent_notification(
         {:subscribe, endpoints},
-        ctx,
-        %{subscribe_mode: :manual} = state
-      ) do
-    subscribed_tracks = Map.keys(state.tracks)
-
-    tracks =
-      state.rtc_engine
-      |> Engine.get_tracks()
-      |> Enum.filter(fn track ->
-        track.origin in endpoints and track.id not in subscribed_tracks
-      end)
-
-    origins = MapSet.new(tracks, fn track -> track.origin end)
-
-    not_found_endpoints =
-      Enum.filter(endpoints, fn endpoint ->
-        MapSet.member?(origins, endpoint)
-      end)
-
-    if not_found_endpoints != [] do
-      Membrane.Logger.info(
-        "Couldn't subscribe on any track from endpoints: #{not_found_endpoints}"
-      )
-    end
-
-    new_subscribed_tracks =
-      tracks
-      |> Enum.map(fn track -> track.id end)
-      |> subscribe_for_tracks(ctx, state)
-
-    new_state =
-      tracks
-      |> add_tracks(new_subscribed_tracks, state)
-      |> Map.update!(:subscribed_endpoints, fn subscribed ->
-        endpoints |> MapSet.new() |> MapSet.union(subscribed)
-      end)
-
-    {[], new_state}
-  end
-
-  @impl true
-  def handle_parent_notification(
-        {:subscribe, _endpoints} = msg,
         _ctx,
-        %{subscribe_mode: :auto} = state
+        state
       ) do
-    Membrane.Logger.warning("""
-    Unexpected message: #{inspect(msg)}.
-    If you want to add tracks manually set `:subscribe_mode` option to `:manual`.
-    """)
-
-    {[], state}
+    subscriber = Subscriber.add_endpoints(endpoints, state.subscriber)
+    {[], %{state | subscriber: subscriber}}
   end
 
   @impl true
@@ -367,7 +312,8 @@ defmodule Membrane.RTC.Engine.Endpoint.HLS do
       end
 
     children_to_remove =
-      state.tracks
+      state.subscriber
+      |> Subscriber.get_tracks()
       |> Enum.flat_map(fn {id, _track} -> Enum.map(@track_children, &{&1, id}) end)
       |> Enum.filter(&Map.has_key?(ctx.children, &1))
 
@@ -381,26 +327,6 @@ defmodule Membrane.RTC.Engine.Endpoint.HLS do
 
   @impl true
   def handle_info(_msg, _ctx, state), do: {[], state}
-
-  defp subscribe_for_tracks(track_ids, ctx, state) do
-    {:endpoint, endpoint_id} = ctx.name
-
-    Enum.filter(track_ids, fn track_id ->
-      case Engine.subscribe(state.rtc_engine, endpoint_id, track_id) do
-        :ok ->
-          true
-
-        :ignored ->
-          false
-      end
-    end)
-  end
-
-  defp add_tracks(tracks, subscribed_tracks, state) do
-    tracks
-    |> Enum.filter(fn track -> track.id in subscribed_tracks end)
-    |> Enum.reduce(state, fn track, state -> put_in(state, [:tracks, track.id], track) end)
-  end
 
   defp get_hls_sink_spec(state, stream_id \\ nil) do
     directory = get_hls_stream_directory(state, stream_id)